Creating NavBar and deleting edge runtime from home page
