+++ package.json
@@ -25,12 +25,14 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "resend": "^4.1.2",
+    "sharp": "^0.33.5",
     "stripe": "^17.4.0-beta.2"
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.13.10",
     "@cloudflare/workers-types": "^4.20250320.0",
     "@eslint/eslintrc": "^3",
+    "@svgr/webpack": "^8.1.0",
     "@tailwindcss/postcss": "^4",
     "@types/js-cookie": "^3.0.6",
     "@types/micro-cors": "^0.1.5",
