+++ src/app/blog/page.tsx
@@ -1,0 +1,7 @@
+export default function Blog() {
+    return (
+        <div>
+            <p>BLOG</p>
+        </div>
+    )
+}
