+++ src/app/globals.css
@@ -14,7 +14,7 @@
 
   --color-blur-500: #F9F9FB4D;
 
-  --color-navbar: #FFFFFF;
+  --color-navbar: #333333;
 
   --color-carrot-50: #FEF3E6;
   --color-carrot-100: #FDE9D3;
