+++ src/app/page.tsx
@@ -1,5 +1,3 @@
-export const runtime = "edge";
-
 import HeroSection from "@/components/HomePage/sections/HeroSection";
 
 import ComparisonSection from "@/components/HomePage/sections/ComparisonSection";
